+++ CHANGES.md
@@ -6,6 +6,8 @@
 
 - ENH: Implementing RADARSAT-Constellation products (as `RCM`)
 - ENH: Adding `RH`, `RV`, `RH_DSPK` and `RV_DSPK` SAR bands
+- ENH: Loading by size -> round resolution to the closest meter (or decimeter for resolution < 1.0m)
+- FIX: Fix band name with complex resolutions
 - FIX: Fixing minor bug in RADARSAT-2 data when looking for product type
 - FIX: Fixing SAR band search in BEAM-DIMAP files
 - FIX: Fixing python version in environment.yml